+++ core/src/main/kotlin/sift/core/api/Action.kt
@@ -86,7 +86,7 @@ import sift.core.terminal.StringEditor
     JsonSubTypes.Type(Action.HasAnnotation::class, name = "has-annotation"),
     JsonSubTypes.Type(Action.EntityFilter::class, name = "entity-filter"),
     JsonSubTypes.Type(Action.FilterModifiers::class, name = "filter-modifiers"),
-    JsonSubTypes.Type(Action.FilterVisibile::class, name = "filter-visible"),
+    JsonSubTypes.Type(Action.FilterVisible::class, name = "filter-visible"),
     JsonSubTypes.Type(Action.ReadAnnotation::class, name = "read-annotation"),
     JsonSubTypes.Type(Action.WithValue::class, name = "with-value"),
     JsonSubTypes.Type(Action.EditText::class, name = "edit-text"),
@@ -889,7 +889,7 @@ sealed class Action<IN, OUT> {
         }
     }
 
-    internal data class FilterVisibile<T : Element>(
+    internal data class FilterVisible<T : Element>(
         val visibility: Visibility,
     ) : IsoAction<T>() {
         override fun id() = "filter-visibility(${visibility.name.lowercase()})"
@@ -897,7 +897,7 @@ sealed class Action<IN, OUT> {
             return input.filter { elem ->
                 visibility == when (elem) {
                     is ClassNode  -> elem.visibility
-                    is FieldNode  -> TODO()
+                    is FieldNode  -> elem.visibility
                     is MethodNode -> elem.visibility
                     else -> error("No access modifiers possible for: ${elem::class.simpleName}")
                 }